requirements.txt
google-adk
streamlit
sqlalchemy
fastapi
uvicorn
pydantic
